+++ types/src/event_info.rs
@@ -8,6 +8,6 @@ pub struct ContractEventInfo {
     pub block_number: BlockNumber,
     pub transaction_hash: HashValue,
     // txn index in block
-    pub transaction_index: u64,
+    pub transaction_index: u32,
     pub event: ContractEvent,
 }
